requirements.txt

asn1crypto==0.24.0
bidict==0.22.1
blinker==1.6.3
cbor2==5.4.6
certifi==2018.8.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.4.0
cryptography==2.6.1
entrypoints==0.3
exceptiongroup==1.2.2
h11==0.14.0
idna==2.6
importlib-metadata==6.7.0
iniconfig==2.0.0
keyring==17.1.1
keyrings.alt==3.1.1
luma.core==2.4.2
luma.oled==3.13.0
packaging==24.0
Pillow==9.5.0
pluggy==1.2.0
pycairo==1.23.0
pycparser==2.21
pycrypto==2.6.1
pycurl==7.43.0.2
pyftdi==0.54.0
PyGObject==3.30.4
pyserial==3.5
pytest==7.4.4
python-engineio==3.14.2
python-socketio==4.6.1
pyusb==1.2.1
pyxdg==0.25
PyYAML==6.0.1
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
simple-websocket==1.1.0
six==1.15.0
smbus2==0.5.0
spidev==3.6
tomli==2.0.1
transitions==0.9.2
typing-extensions==4.7.1
urllib3==1.24.1
websocket-client==1.6.1
wsproto==1.2.0
zipp==3.15.0


